Make it possible to add an object to value
In our menu we want some of the buttons to eighter navigate to internal or external url's. We pass the url to the MenuItem using the value property, so onSelection can trigger the navigation. The problem is that we use React Router which requires us to either call Router.transitionTo() or window.location depending on if the URL is internal or external. We cant t determine if an url is an internal since the value is just a string. If the value could be an object, we could send { internal:true, url:'http://url' }